Refactor CronParser class and split file parsing related logic into CronFileParser class
