+++ src/index.ts
@@ -1,11 +1,11 @@
 /* istanbul ignore file */
-export { CronFieldCollection } from './CronFieldCollection';
-export { CronParser } from './CronParser';
+import { CronExpressionParser } from './CronExpressionParser';
+
 export { CronDate } from './CronDate';
+export { CronFieldCollection } from './CronFieldCollection';
 export { CronExpression, CronExpressionOptions } from './CronExpression';
 export { CronExpressionParser } from './CronExpressionParser';
-import { CronParser } from './CronParser';
-
-export default CronParser;
+export { CronFileParser, CronFileParserResult } from './CronFileParser';
 
 export * from './fields';
+export default CronExpressionParser;
