+++ tests/CronExpressionParser.test.ts
@@ -1515,68 +1515,138 @@ describe('CronExpressionParser', () => {
     });
   });
 
-  // describe('parse', () => {
-  //   test('parse a expression correctly', () => {
-  //     const expression = '* * * * *';
-  //     const cronExpression = CronExpressionParser.parse(expression);
-  //     expect(cronExpression).toBeInstanceOf(CronExpression);
-  //   });
+  describe('test expressions with "L" last of flag', () => {
+    const testCasesLastWeekdayOfMonth = [
+      { expression: '0 0 0 * * 1L', expectedDate: 27 },
+      { expression: '0 0 0 * * 2L', expectedDate: 28 },
+      { expression: '0 0 0 * * 3L', expectedDate: 29 },
+      { expression: '0 0 0 * * 4L', expectedDate: 30 },
+      { expression: '0 0 0 * * 5L', expectedDate: 24 },
+      { expression: '0 0 0 * * 6L', expectedDate: 25 },
+      { expression: '0 0 0 * * 0L', expectedDate: 26 },
+      { expression: '0 0 0 * * 7L', expectedDate: 26 },
+    ];
 
-  //   test('parse predefined expressions correctly', () => {
-  //     Object.entries(PredefinedExpressions).forEach(([key, value]) => {
-  //       const cronExpression = CronExpressionParser.parse(key);
-  //       expect(cronExpression.stringify(true)).toBe(value);
-  //     });
-  //   });
+    test('parse cron with last day in a month', () => {
+      const options = {
+        currentDate: new Date(2014, 0, 1),
+        endDate: new Date(2014, 10, 1),
+      };
 
-  //   test('throw an error when using both dayOfMonth and dayOfWeek together in strict mode', () => {
-  //     const expression = '0 0 * * 1-5';
-  //     expect(() => CronExpressionParser.parse(expression, { strict: true })).toThrow();
-  //   });
+      const interval = CronExpressionParser.parse('0 0 L * *', options);
+      expect(interval.hasNext()).toBe(true);
 
-  //   test('parse expressions with aliases correctly', () => {
-  //     const expression = '0 0 1 JAN SUN';
-  //     const cronExpression = CronExpressionParser.parse(expression);
-  //     expect(cronExpression.stringify()).toBe('0 0 1 1 0');
-  //   });
+      for (let i = 0; i < 10; ++i) {
+        const next = interval.next();
+        expect(next).toBeDefined();
+      }
+    });
 
-  //   test('throw an error when an invalid character is encountered', () => {
-  //     const expression = '0 0 1 JAN SUNDAY';
-  //     expect(() => CronExpressionParser.parse(expression)).toThrow();
-  //   });
+    test('parse cron with last day in feb', () => {
+      const options = {
+        currentDate: new Date(2016, 0, 1),
+        endDate: new Date(2016, 10, 1),
+      };
 
-  //   test('parse expressions with L correctly', () => {
-  //     const expression = '0 0 L * *';
-  //     const cronExpression = CronExpressionParser.parse(expression);
-  //     expect(cronExpression.stringify()).toBe('0 0 L * *');
-  //   });
+      const interval = CronExpressionParser.parse('0 0 6-20/2,L 2 *', options);
+      expect(interval.hasNext()).toBe(true);
 
-  //   test('throw an error when an invalid range is encountered', () => {
-  //     const expression = '0 0 32 * *';
-  //     expect(() => CronExpressionParser.parse(expression)).toThrow();
-  //   });
-  // });
+      const items = 9;
+      let next;
+      let i = 0;
+      while (interval.hasNext()) {
+        next = interval.next();
+        i += 1;
+        expect(next).toBeDefined();
+      }
 
-  // describe('stringify and debug', () => {
-  //   test('return the correct string representation and debug object', () => {
-  //     const expression = '0 0 1,15 * 1';
-  //     const cronExpression = CronExpressionParser.parse(expression);
-  //     expect(cronExpression.stringify()).toBe('0 0 1,15 * 1');
-  //   });
+      //leap year
+      if (!next) {
+        throw new Error('Invalid date');
+      }
+      expect(next).not.toBeUndefined();
+      expect(next.getDate()).toBe(29);
+      expect(i).toBe(items);
+    });
 
-  //   test('handle complex expressions correctly', () => {
-  //     const expression = '*/15 10-20,22 1-7,15,31 1,3,5,7-11 1-5';
-  //     const cronExpression = CronExpressionParser.parse(expression);
-  //     expect(cronExpression.stringify()).toBe('*/15 10-20,22 1-7,15,31 1,3,5,7,8-11 1-5');
-  //   });
+    test('parse cron with last day in feb', () => {
+      const options = {
+        currentDate: new Date(2014, 0, 1),
+        endDate: new Date(2014, 10, 1),
+      };
 
-  //   test('throw an error when an invalid character is encountered in a specific field', () => {
-  //     expect(() => CronExpressionParser.parse('65 * * * *')).toThrow();
-  //     expect(() => CronExpressionParser.parse('* 70 * * *')).toThrow();
-  //     expect(() => CronExpressionParser.parse('* * 32 * *')).toThrow();
-  //     expect(() => CronExpressionParser.parse('* * * 14 *')).toThrow();
-  //     expect(() => CronExpressionParser.parse('* * * * 8')).toThrow();
-  //   });
-  // });
+      const interval = CronExpressionParser.parse('0 0 1,3,6-10,L 2 *', options);
+      expect(interval.hasNext()).toBe(true);
+
+      let next;
+      while (interval.hasNext()) {
+        next = interval.next();
+        expect(next).toBeDefined();
+      }
+      if (!next) {
+        throw new Error('Invalid date');
+      }
+      expect(next).not.toBeUndefined();
+      expect(next.getDate()).toBe(28);
+    });
+
+    testCasesLastWeekdayOfMonth.forEach(({ expression, expectedDate }) => {
+      const options = {
+        currentDate: new Date(2021, 8, 1),
+        endDate: new Date(2021, 11, 1),
+      };
+
+      test(`parse cron with last weekday of the month: ${expression}`, () => {
+        const interval = CronExpressionParser.parse(expression, options);
+        expect(interval.hasNext()).toBe(true);
+        const next = interval.next();
+        expect(next.getDate()).toBe(expectedDate);
+      });
+    });
+
+    test('parses expression that runs on both last monday and friday of the month', () => {
+      const options = {
+        currentDate: new Date(2021, 8, 1),
+        endDate: new Date(2021, 11, 1),
+      };
+      const interval = CronExpressionParser.parse('0 0 0 * * 1L,5L', options);
+      let next = interval.next();
+
+      expect(next.getDate()).toBe(24);
+      next = interval.next();
+
+      expect(next.getDate()).toBe(27);
+    });
+
+    test('parses expression that runs on both every monday and last friday of month', () => {
+      const options = {
+        currentDate: new Date(2021, 8, 1),
+        endDate: new Date(2021, 8, 30),
+      };
+      const interval = CronExpressionParser.parse('0 0 0 * * 1,5L', options);
+      const dates: number[] = [];
+      let isNotDone = true;
+      while (isNotDone) {
+        try {
+          const next = interval.next();
+          dates.push(next.getDate());
+        } catch (e) {
+          if (e instanceof Error && e.message !== 'Out of the timespan range') {
+            throw e;
+          }
+          isNotDone = false;
+          break;
+        }
+      }
+      expect(dates).toEqual([6, 13, 20, 24, 27]);
+    });
+
+    test('throw new Errors to parse for invalid last weekday of month expression', () => {
+      expect(() => {
+        const interval = CronExpressionParser.parse('0 0 0 * * L');
+        interval.next();
+      }).toThrow();
+    });
+  });
 });
